+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## [0.2.0](https://github.com/typescript-tea/core/compare/v0.1.0...v0.2.0) - 2020-02-26
+
+### Changed
+
+- Renamed runtime to Program.run()
+- Renamed mapDispatch to Dispatch.map()
+- Require render function to Program.run()
+- Make effectManagers optional in Program.run()
+
 ## [0.1.0](https://github.com/typescript-tea/core/compare/v0.1.0...v0.1.0) - 2020-01-26
 
 ### Added